refactor: 공지사항 구현 UI index설정 및 style수정
