+++ src/router/index.js
@@ -58,15 +58,13 @@ const router = createRouter({
             component: () => import('../pages/AdminPage.vue'),
             children: [
                 {
-                    path: 'notice',        // 공지사항 리스트
+                    path: 'notice', 
                     component: () => import('../components/admin/NoticeAdmin.vue')
-                },
-                // {
-                //     path: 'notice/:id',     // 공지사항 상세페이지
-                //     component: () => import('../components/admin/NoticeDetail.vue'),
-                //     props: true
-                // }
-            ]
+                }
+            ]},
+            {
+                path: '/adminPage/notice/:id',
+                component: () => import('../components/admin/NoticeContentPage.vue')
         },
         {
             path: '/findid/result',
